requirements.txt
fastapi
uvicorn
python-multipart
inference-sdk
opencv-python
numpy
matplotlib
Pillow
pydantic
python-dotenv
sqlalchemy
python-jose
alembic
psycopg2-binary
aiosqlite
pytest
httpx
selenium
webdriver-manager
pytest-cov
uvicorn 